+++ NEWS.md
@@ -9,6 +9,8 @@
 * Provides `holdings-storage 8.1`
 * Provides `instance-storage 11.2`
 * Provides `service-points 3.5`
+* Provides `inventory-hierarchy 0.7`
+* Provides `locations 3.2`
 
 ### Features
 * Make max.request.size configurable for reindex holdings/items producers ([MODINVSTOR-1372](https://folio-org.atlassian.net/browse/MODINVSTOR-1372))
@@ -21,6 +23,10 @@
 * New enum property `defaultCheckInActionForUseAtLocation` for reading room circulation ([MODINVSTOR-1376](https://folio-org.atlassian.net/browse/MODINVSTOR-1376))
 * Holdings record statistical codes validation enhancement ([MODINVSTOR-1185](https://folio-org.atlassian.net/browse/MODINVSTOR-1185))
 * Adds an array to hold additional call numbers on the holding and item level ([MODINVSTOR-1432](https://folio-org.atlassian.net/browse/MODINVSTOR-1432))
+* Update `get_items_and_holdings_view` to properly handle inactive locations ([MODINVSTOR-1438](https://folio-org.atlassian.net/browse/MODINVSTOR-1438))
+* Update get_items_and_holdings_view to get location status ([MODINVSTOR-1447](https://folio-org.atlassian.net/browse/MODINVSTOR-1447))
+* Remove migration script to update 'deleted' field of instance.jsonb ([MODINVSTOR-1449](https://folio-org.atlassian.net/browse/MODINVSTOR-1449))
+* Add new `isShadow` property to location schema ([MODINVNSTOR-1436](https://folio-org.atlassian.net/browse/MODINVSTOR-1436))
 
 ### Bug fixes
 * Fix ordering of electronic access items for inventory-hierarchy, oai-pmh-view ([MODINVSTOR-1224](https://folio-org.atlassian.net/browse/MODINVSTOR-1224))
@@ -40,6 +46,8 @@
 * Stabilize tests by fixing  cache issue during create/update holding ([MODINVSTOR-1405](https://folio-org.atlassian.net/browse/MODINVSTOR-1405))
 * Refactor usage of deprecated `PostgresClient` methods ([MODINVSTOR-1367](https://folio-org.atlassian.net/browse/MODINVSTOR-1367))
 * Refactor controllers to extract common logic into shared class ([MODINVSTOR-1437](https://folio-org.atlassian.net/browse/MODINVSTOR-1437))
+* Cover contributor-types, contributor-name-types, instance-formats, instance-types, nature-of-content-terms, classification-types, modes-of-issuance APIs with integration tests ([MODINVSTOR-1166](https://folio-org.atlassian.net/browse/MODINVSTOR-1166))
+* Cover instance-statuses, electronic-access-relationships, statistical-code-types, statistical-codes, ill-policies, holdings-types, call-number-types APIs with integration tests ([MODINVSTOR-1167](https://folio-org.atlassian.net/browse/MODINVSTOR-1167))
 
 ### Dependencies
 * Bump `LIB_NAME` from `OLD_VERSION` to `NEW_VERSION`